requirements/base.txt
# Core Django and Web Framework
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3.0

# Database and ORM
psycopg2-binary>=2.9.7
django-extensions>=3.2.0

# Caching and Task Queue
redis>=4.6.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# AI and Content Generation
openai>=1.0.0
anthropic>=0.3.0
langchain>=0.0.300
tiktoken>=0.5.0

# Web Scraping and Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.10.0
selenium>=4.15.0
PyGithub>=1.59.0
feedparser>=6.0.10

# Text Processing and Analysis
spacy>=3.7.0
nltk>=3.8.1
textstat>=0.7.0
# Text Processing
Markdown>=3.5.0
markdown2>=2.4.0

# Image Processing
Pillow>=10.0.0
django-imagekit>=4.1.0

# Authentication and Security
django-allauth>=0.57.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Email and Communications
sendgrid>=6.10.0
django-ses>=3.5.0

# Static Files and Storage
django-storages>=1.14.0
boto3>=1.28.0
django-compressor>=4.4.0
whitenoise>=6.5.0

# SEO and Meta Management
django-meta>=2.2.0
django-sitemap>=2.2.0

# Configuration Management
python-decouple>=3.8
django-environ>=0.11.0

# Monitoring and Logging
sentry-sdk>=1.32.0
structlog>=23.0.0

# Performance and Utilities
django-debug-toolbar>=4.2.0
django-health-check>=3.17.0
django-widget-tweaks>=1.4.12
django-crispy-forms>=2.0

# API Documentation
drf-spectacular>=0.26.0

# Date and Time
python-dateutil>=2.8.2
pytz>=2023.3
